+++ assess/src/sdk/sdk.ts
@@ -51,8 +51,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0";
-    sdkVersion = "0.31.2";
-    genVersion = "2.86.10";
+    sdkVersion = "0.31.3";
+    genVersion = "2.87.1";
 
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);